prepare architecture for tests
