+++ meilisearch-http/src/main.rs
@@ -1,13 +1,12 @@
 use std::{env, thread};
 
 use actix_cors::Cors;
-use actix_web::{middleware, web, App, HttpServer};
+use actix_web::{middleware, HttpServer};
 use log::info;
 use main_error::MainError;
 use meilisearch_http::data::Data;
 use meilisearch_http::option::Opt;
-use meilisearch_http::routes;
-use meilisearch_http::routes::index_update_callback;
+use meilisearch_http::{create_app, index_update_callback};
 use structopt::StructOpt;
 
 mod analytics;
@@ -51,7 +50,7 @@ async fn main() -> Result<(), MainError> {
     print_launch_resume(&opt, &data);
 
     HttpServer::new(move || {
-        App::new()
+        create_app(&data)
             .wrap(
                 Cors::new()
                     .send_wildcard()
@@ -60,59 +59,6 @@ async fn main() -> Result<(), MainError> {
             )
             .wrap(middleware::Logger::default())
             .wrap(middleware::Compress::default())
-            .app_data(web::Data::new(data.clone()))
-            .wrap(routes::Authentication::Public)
-            .service(routes::load_html)
-            .service(routes::load_css)
-            .service(routes::search::search_with_url_query)
-            .service(routes::search::search_multi_index)
-            .service(routes::document::get_document)
-            .service(routes::document::get_all_documents)
-            .wrap(routes::Authentication::Private)
-            .service(routes::index::list_indexes)
-            .service(routes::index::get_index)
-            .service(routes::index::create_index)
-            .service(routes::index::update_index)
-            .service(routes::index::delete_index)
-            .service(routes::index::get_update_status)
-            .service(routes::index::get_all_updates_status)
-            .service(routes::document::delete_document)
-            .service(routes::document::add_documents)
-            .service(routes::document::update_documents)
-            .service(routes::document::delete_documents)
-            .service(routes::document::clear_all_documents)
-            .service(routes::setting::update_all)
-            .service(routes::setting::get_all)
-            .service(routes::setting::delete_all)
-            .service(routes::setting::get_rules)
-            .service(routes::setting::update_rules)
-            .service(routes::setting::delete_rules)
-            .service(routes::setting::get_distinct)
-            .service(routes::setting::update_distinct)
-            .service(routes::setting::delete_distinct)
-            .service(routes::setting::get_searchable)
-            .service(routes::setting::update_searchable)
-            .service(routes::setting::delete_searchable)
-            .service(routes::setting::get_displayed)
-            .service(routes::setting::update_displayed)
-            .service(routes::setting::delete_displayed)
-            .service(routes::setting::get_accept_new_fields)
-            .service(routes::setting::update_accept_new_fields)
-            .service(routes::stop_words::get)
-            .service(routes::stop_words::update)
-            .service(routes::stop_words::delete)
-            .service(routes::synonym::get)
-            .service(routes::synonym::update)
-            .service(routes::synonym::delete)
-            .service(routes::stats::index_stats)
-            .service(routes::stats::get_stats)
-            .service(routes::stats::get_version)
-            .service(routes::stats::get_sys_info)
-            .service(routes::stats::get_sys_info_pretty)
-            .service(routes::health::get_health)
-            .service(routes::health::change_healthyness)
-            .wrap(routes::Authentication::Admin)
-            .service(routes::key::list)
     })
     .bind(opt.http_addr)?
     .run()
